requirements.txt
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
idna==3.10
isodate==0.7.2
lxml==6.0.1
onvif_zeep==0.2.12
pillow==11.3.0
platformdirs==4.4.0
pytz==2025.2
requests==2.32.5
requests-file==2.1.0
requests-toolbelt==1.0.0
urllib3==2.5.0
zeep==4.3.1
